Moved cache databases type into unique namespace
