+++ GroupMeClient/GroupMeClient.csproj
@@ -77,6 +77,9 @@
     <Compile Include="Caching\Migrations\20200630021510_StarMessageSupport.designer.cs">
       <DependentUpon>20200630021510_StarMessageSupport.cs</DependentUpon>
     </Compile>
+    <Compile Include="Caching\Models\GroupIndexStatus.cs" />
+    <Compile Include="Caching\Models\IDatabaseMessageWrapper.cs" />
+    <Compile Include="Caching\Models\StarredMessage.cs" />
     <Compile Include="Caching\SuperIndexer.cs" />
     <Compile Include="Converters\IsGreaterThanConverter.cs" />
     <Compile Include="Converters\PercentageConverter.cs" />
